+++ tests/testthat/test-python-backend.R
@@ -527,8 +527,7 @@ test_that("stored deps versions translate to py_require() arguments", {
     args$packages,
     c(
       "tensorflow==2.14.*",
-      "tensorflow_probability==0.22.*",
-      "tf-keras==2.14.*"
+      "tensorflow_probability[tf]==0.22.*"
     )
   )
   expect_equal(args$python_version, "3.10")